FPS-camera movement now smoother. Also considered as fix for Bug #358. ICursorControl::getPosition and ICursorControl::getRelativePosition now have a parameter updateCursor. When that is set to false the last set cursor-position is returned without asking OS for a new one.
The problem was that FPS camera does call setRelativePosition(0.5, 0.f) followed by getRelativePosition to get the values which have been set.
We need that part because due to int/float conversions (mouse-positions are in pixels) and maybe also sometimes floating point inaccuracies the center values
after setting it is often no longer 0.5. But this did not always just return the 0.5 converted to screen coordinates, but often rather different values. 
Not sure about the exact problem (debugging cursor-positions is tricky...), but it was either
- the mouse-position could change between those 2 calls (my main guess, thought it doesn't explain why it then seemed to move camera backward as we reset to whatever position we get, so it should only lose a mouse-movement) 
- the new position was not yet reached and we got the old position (debugging returned positions which only seemed close to old position)
- there is some interpolation going on sometimes in the OS and we got a position between the 0.5 we set and the old position the cursor had (don't thinks so, but would explain the camera moving backwards).
Whatever the reason, this somehow caused the fps-camera to sometimes rotate in the wrong direction for a frame. Hardly noticable with good framerates, but with some bad framerates this caused to camera sometimes to jump around a lot.


git-svn-id: svn://svn.code.sf.net/p/irrlicht/code/trunk@5950 dfc29bdd-3216-0410-991c-e03cc46cb475