Fix fps camera animator losing mouse-events and switch input to real-time.
Real-time instead of virtual time because it's about user-input and users are acting in real-time.

The bigger change is that the fps-animator lost all mouse-events between device->run() and it's animateNode() call.
It checked the cursor-position only on mouse-move events. But did reset the cursor in animateNode. That is called often many milliseconds after events have been handled. 
All cursor movement in between got lost. 
Additionally this caused getRelativePosition to be called far more often than necessary (which can be rather expensive on X11).
The result was that the fps camera jumped around hard with bad framerates and even with good framerates it was rotating slower.
I tried fixing fps-cam troubles already recently in r5950, but while that improved it a tiny bit I had still missed the real big problems back then. 
FPS-Camera is now far smoother, especially with bad FPS it's a day & night difference.



git-svn-id: svn://svn.code.sf.net/p/irrlicht/code/trunk@5965 dfc29bdd-3216-0410-991c-e03cc46cb475